requirements.txt
# Core LangChain
langchain
langchain-core
langchain-openai
langchain-groq
langchain-text-splitters
langchain_community
langchain_chroma 

# Document Loaders
pypdf
pypdf2
beautifulsoup4
requests

# Tracing & Monitoring
langsmith
langgraph

# Web Framework
fastapi
uvicorn
python-multipart
aiofiles

# Database & Auth
sqlalchemy
passlib
bcrypt==4.0.1  # Kept version as requested
pyjwt
python-multipart
python-jose[cryptography]

# Utilities
python-dotenv
pydantic
docx2txt
streamlit
google-search-results
ipykernel
ipywidgets

sentence-transformers
langchain-huggingface
pytesseract 
pillow 
pdf2image
addict 
matplotlib 
torchvision
einops
easydict
docx



torch==2.6.0
transformers==4.46.3
tokenizers==0.20.3
#flash-attn==2.7.3 

